backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
markitdown==0.0.1a2
python-multipart==0.0.6
pillow==10.1.0
werkzeug==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23